파일 업로드 기능 구현 FFB-2

+++ src/main/java/com/fivefeeling/memory/dto/PinPointSummaryDTO.java
@@ -3,8 +3,8 @@ package com.fivefeeling.memory.dto;
 public record PinPointSummaryDTO(
     Long tripId,
     Long pinPointId,
-    Float latitude,
-    Float longitude
+    Double latitude,
+    Double longitude
 ) {
 
 }
